requirements.txt
adb-shell==0.4.3
Appium-Python-Client==1.0.2
attrs==23.1.0
cffi==1.15.1
cryptography==41.0.1
iniconfig==2.0.0
mock==5.0.2
packaging==23.1
pluggy==0.13.1
pure-python-adb==0.3.0.dev0
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
pytest==6.1.1
python-dotenv==1.0.0
rsa==4.9
selenium==3.141.0
six==1.16.0
sure==2.0.1
toml==0.10.2
urllib3==1.26.5
